Rename to ArrayView to better specify its purpose.
