+++ catkit_core/MessageBroker.cpp
@@ -517,17 +517,23 @@ void MessageBroker::PublishData(std::string_view topic, const void *data, size_t
 	PublishMessage(message);
 }
 
-void MessageBroker::PublishData(std::string_view topic, const Tensor &tensor, uint8_t memory_block_id)
+void MessageBroker::PublishData(std::string_view topic, const ArrayView &array, uint8_t memory_block_id)
 {
 	Uuid trace_id;
 	Uuid::Generate(&trace_id);
 
-	PublishData(topic, tensor, trace_id, memory_block_id);
+	PublishData(topic, array, trace_id, memory_block_id);
 }
 
-void MessageBroker::PublishData(std::string_view topic, const Tensor &tensor, Uuid trace_id, uint8_t memory_block_id)
+void MessageBroker::PublishData(std::string_view topic, const ArrayView &array, Uuid trace_id, uint8_t memory_block_id)
 {
-	// TODO.
+	auto message = PrepareMessage(topic, array.info.GetSizeInBytes(), trace_id, memory_block_id);
+
+	// Copy over array and array info.
+	std::memcpy(message.m_Payload, array.data, array.info.GetSizeInBytes());
+	message.SetArrayInfo(array.info);
+
+	PublishMessage(message);
 }
 
 std::optional<Message> MessageBroker::TryGetMessage(std::string_view topic, size_t frame_id)
@@ -744,9 +750,9 @@ void Message::SetArrayInfo(const ArrayInfo &array_info)
 	m_Header->payload_info.array_info = array_info;
 }
 
-void *Message::GetPayload() const
+ArrayView Message::GetPayload() const
 {
-	return m_Payload;
+	return ArrayView(m_Header->payload_info.array_info, m_Payload);
 }
 
 std::size_t Message::GetPayloadSize() const
